requirements.txt
streamlit
Pillow
requests
opencv-python
numpy
supabase
python-dotenv
opencv-python